--- a/src/functor.jl
+++ b/src/functor.jl
@@ -169,7 +169,9 @@ function gpu(x)
   if hasmethod(_gpu, Tuple{Any})
     _gpu(x)
   else
-    error("CUDA not loaded. Load `CUDA` and `NNlibCUDA` to access GPU functionality")
+    @info """The GPU functionality is being called via `Flux.gpu` but `CUDA` and `NNlibCUDA` \
+    must be loaded to access GPU functionality""" maxlog=1
+    x
   end
 end