@IanButterworth 2022-12-09
One question, I structured this so that you need to load both `CUDA` and `NNlibCUDA`, which means you can avoid downloading & Pkg.precompiling both of those. 

Alternatively `NNlibCUDA` could be made a strong dep so you just have to `using CUDA`, and it could be loaded in the extension