@DhairyaLGandhi 2022-12-09
I prefer the latter since it enables a nicer user experience. Does the added precompilation cost a tonne of time?