@ToucheSir 2022-12-09
Is there a way we could have CUDA.jl as the only trigger dep? Up to this point we've been acting on the model that users shouldn't have to know NNlibCUDA even exists, so if possible it would be great to keep that.