@IanButterworth 2022-12-09
@KristofferC would that be possible? NNlibCUDA deps on CUDA so AFAIK it's not possible