@ToucheSir 2022-12-09
If needed, any and all changes required on the NNlibCUDA side should be on the table. Would making its CUDA dependency also weak be enough?