@IanButterworth 2022-12-09
Maybe absorb it into NNlib and make CUDA a weak dep of NNlib?