@KristofferC 2022-12-09
Note that Registrator is not updated yet to handle weak deps.