@CarloLucibello 2022-12-09
Two suggestions, although the first one is likely unrealizable:
1) Can `gpu` automatically load the extension if `CUDA` is available in the environment (although not loaded)?

2) If the answer to 1) is not, I would prefer we have a registered (sub)package named `FluxCUDA` which loads the extension, and the warning points to that. One should also be able to do `using FluxCUDA` directly instead of `using Flux, FluxCUDA` and get all the functionality.