@IanButterworth 2022-12-09
> 1. Can gpu automatically load the extension if CUDA is available in the environment (although not loaded)?

I haven't thought through the implications of this design, but yes it should be possible via something like
```
Base.locate_package(Base.PkgId("NNlibCUDA", {uuid}) !== nothing && @eval Main using NNlibCUDA
```

There's one test failure remaining that I don't understand:
```
Dropout RNGs: Test Failed at /var/lib/buildkite-agent/builds/gpuci-17/julialang/flux-dot-jl/test/cuda/layers.jl:285
--
  | Expression: Flux.dropout(MersenneTwister(), CUDA.rand(Float32, 2, 3), 0.1)
  | Expected: ArgumentError
  | No exception thrown
```