@CarloLucibello 2022-12-09
> I haven't thought through the implications of this design, but yes it should be possible via something like

Wouldn't that incur in world age issues? That is the limitation of the approach in [LazyModules.jl](https://github.com/johnnychen94/LazyModules.jl#world-age-issue)