@KristofferC 2022-12-09
It would be good to have a concrete usage example here. Like maybe a toy session in a REPL to see how a user would use it. And then based on that it might be easier to come up with a good design.