@CarloLucibello 2022-12-09
Ideally, the following could happen, which is also non-breaking:
```julia
julia> using Flux  # only cpu code imported, CUDA.jl and NNlibCUDA.jl are not loaded yet

julia> x = [1.0,2.0,3.0];
3-element Vector{Float64}:
 1.0
 2.0
 3.0

julia> gpu(x) # here the gpu code and libraries are loaded if the device has cuda capabilities            
3-element CUDA.CuArray{Float32, 1, CUDA.Mem.DeviceBuffer}:
 1.0
 2.0
 3.0

# successive calls to `gpu` are fast
```