@ToucheSir 2022-12-09
Even something like the current behaviour of warning and passing the value through when CUDA isn't loaded could work, IMO.