@mcabbott 2022-12-10
Locally, it looks like I can have just `CUDAExt = "CUDA"`, and CUDAExt is able to `import NNlibCUDA` and call functions from it, or other packages which I load only within the extension. But I didn't test very carefully, does this fail for some reason?