@mcabbott 2022-12-10
Yes, but I don't know if it matters. It must remain in [deps] for now I think.