@IanButterworth 2022-12-10
Well that means it's not a weak dep, so it and CUDA will be downloaded, installed and precompiled. You do avoid loading it until the extension loads, but personally I'm motivated by both avoiding installing CUDA etc and loading it.