@IanButterworth 2022-12-10
For more context, using Flux on a resource constrained system like a small embedded SBC without a gpu is painful because of all of the above, not just load time