@mcabbott 2022-12-10
So a package listed under both deps and weakdeps (as in the present state of this PR), is effectively deleted from deps.

I'm a bit surprised that the design doesn't allow for loading 3rd packages, but maybe this is hard, didn't follow closely. 

Not downloading would be nice too. But requiring anyone to know about this obscure thing called NNlibCUDA seems like a step backwards. Moving it to be an extension of NNlib is probably the way to go then.