@CarloLucibello 2022-12-10
Having NNlibCUDA as an extension of NNlib seems the way to go. Also, we finally move back that code to the original repo as we wanted to do for some time. Since `NNlibCUDA` was not part of `NNlib` I guess we should have
```julia
# In NNlib.jl
if !isdefined(Base, :get_extension)
  # do nothing
end
```
and continue to import `NNlibCUDA` on old julia versions of Flux.