@mcabbott 2022-12-10
So the proposal is to have NNlibCUDA under both [deps] and [weakdeps] here, load it unconditionally on Julia 1.6, but don't try to load it at all on 1.10. 

Then give NNlib an extension for CUDA, which has another copy of the exact same code as the package NNlibCUDA? If the registered package is moved to the same repository as NNlib, then both the package and NNlib's CUDAext can `include` the same source file.