@ToucheSir 2022-12-10
https://github.com/FluxML/NNlib.jl/pull/445 is now up for anyone who wants to kick the tires with this. I haven't tested using times or TTFX, so any data would be appreciated.