@mcabbott 2022-12-11
BTW, timing this (on a slow machine which has a GPU, Julia nightly):
```julia
julia> @time using Flux
 21.362717 seconds (24.70 M allocations: 1.449 GiB, 5.16% gc time, 46.46% compilation time: 61% of which was recompilation)  # before
  9.096614 seconds (5.34 M allocations: 373.462 MiB, 4.08% gc time, 81.04% compilation time: 56% of which was recompilation)  # after
```
or, loading everything:
```julia
julia> @time using Flux, NNlibCUDA, CUDA
 21.134435 seconds (24.76 M allocations: 1.453 GiB, 3.64% gc time, 47.75% compilation time: 62% of which was recompilation)  # after
```

After this, the biggest offenders in `@time_imports` are FoldsThreads & StaticArrays, details below.

<details>

Before:

```julia
julia> @time_imports using Flux
      1.2 ms  Statistics
     16.7 ms  MacroTools
      0.2 ms  Reexport
      6.8 ms  ProgressLogging
      8.7 ms  IrrationalConstants
      0.2 ms  Compat
    138.7 ms  ChainRulesCore
      8.1 ms  DocStringExtensions 80.65% compilation time
      1.1 ms  ChangesOfVariables
      1.4 ms  InverseFunctions
      1.2 ms  LogExpFunctions
      0.2 ms  OpenLibm_jll
     31.0 ms  Preferences
      0.3 ms  JLLWrappers
    461.8 ms  OpenSpecFun_jll 115.01% compilation time (86% recompilation)
     44.9 ms  SpecialFunctions
      0.3 ms  Requires
      0.3 ms  Adapt
     56.2 ms  NNlib 62.75% compilation time (12% recompilation)
     13.8 ms  ShowCases
      1.8 ms  ConstructionBase
    180.8 ms  InitialValues
      0.2 ms  DataValueInterfaces
      1.3 ms  DataAPI
      0.1 ms  IteratorInterfaceExtensions
      0.1 ms  TableTraits
     13.4 ms  OrderedCollections
     25.2 ms  Tables
      0.2 ms  ZygoteRules
      4.7 ms  StaticArraysCore
     19.4 ms  Setfield
     87.5 ms  BangBang 58.56% compilation time
      1.6 ms  ContextVariablesX
      0.1 ms  FLoopsBase
      1.3 ms  PrettyPrint
      0.5 ms  NameResolution
     31.8 ms  MLStyle
      2.4 ms  JuliaVariables
      0.5 ms  ArgCheck
     22.5 ms  Baselet
      0.1 ms  CompositionsBase
      0.1 ms  DefineSingletons
     17.1 ms  MicroCollections
     16.4 ms  SplittablesBase
    116.1 ms  Transducers 36.62% compilation time
      7.6 ms  FLoops
     41.7 ms  Accessors 39.28% compilation time
     22.2 ms  FunctionWrappers
   1017.0 ms  FoldsThreads 193.39% compilation time
    161.1 ms  DataStructures
      0.5 ms  SortingAlgorithms
     73.7 ms  Missings
      0.4 ms  StatsAPI
     51.0 ms  StatsBase
      5.5 ms  SimpleTraits
      1.0 ms  DelimitedFiles
     11.3 ms  MLUtils
      5.0 ms  Functors
     16.4 ms  Optimisers
      2.4 ms  GPUArraysCore
      0.1 ms  RealDot
     46.6 ms  StructArrays
     37.5 ms  ChainRules
     20.1 ms  IRTools
      0.8 ms  DiffRules
      0.3 ms  NaNMath
    420.6 ms  FillArrays
     20.7 ms  AbstractFFTs
      4.2 ms  DiffResults
   1271.6 ms  StaticArrays
      0.5 ms  CommonSubexpressions
    198.6 ms  ForwardDiff
    174.3 ms  Zygote 52.62% compilation time
      8.9 ms  CEnum
    474.3 ms  LLVMExtra_jll 117.69% compilation time (84% recompilation)
    406.2 ms  LLVM 35.54% compilation time
      0.4 ms  ExprTools
    180.9 ms  TimerOutputs 19.49% compilation time
    823.8 ms  GPUCompiler 22.62% compilation time (68% recompilation)
   1797.9 ms  GPUArrays
     13.6 ms  BFloat16s
     47.7 ms  RandomNumbers 37.30% compilation time
     12.4 ms  Random123
   6794.8 ms  CUDA 2.02% compilation time
     84.9 ms  OneHotArrays
    136.9 ms  NNlibCUDA
    100.9 ms  Flux
```

After:

```julia
julia> @time_imports using Flux
      1.2 ms  Statistics
     41.7 ms  MacroTools
      0.2 ms  Reexport
      6.7 ms  ProgressLogging
      8.5 ms  IrrationalConstants
      0.1 ms  Compat
    136.0 ms  ChainRulesCore
      8.2 ms  DocStringExtensions 82.53% compilation time
      1.1 ms  ChangesOfVariables
      1.3 ms  InverseFunctions
      1.1 ms  LogExpFunctions
      0.2 ms  OpenLibm_jll
     29.5 ms  Preferences
      0.3 ms  JLLWrappers
    453.6 ms  OpenSpecFun_jll 113.71% compilation time (87% recompilation)
     31.0 ms  SpecialFunctions
      0.3 ms  Requires
      0.3 ms  Adapt
     54.2 ms  NNlib 66.10% compilation time (12% recompilation)
     13.3 ms  ShowCases
      1.8 ms  ConstructionBase
     31.1 ms  InitialValues
      0.1 ms  DataValueInterfaces
      1.3 ms  DataAPI
      0.1 ms  IteratorInterfaceExtensions
      0.1 ms  TableTraits
     14.2 ms  OrderedCollections
     25.8 ms  Tables
      0.2 ms  ZygoteRules
      4.9 ms  StaticArraysCore
     20.1 ms  Setfield
     89.9 ms  BangBang 57.72% compilation time
      1.6 ms  ContextVariablesX
      0.1 ms  FLoopsBase
      1.3 ms  PrettyPrint
      0.4 ms  NameResolution
     33.0 ms  MLStyle
      2.5 ms  JuliaVariables
      0.5 ms  ArgCheck
     23.0 ms  Baselet
      0.1 ms  CompositionsBase
      0.1 ms  DefineSingletons
     17.6 ms  MicroCollections
     17.3 ms  SplittablesBase
    117.3 ms  Transducers 36.25% compilation time
      7.8 ms  FLoops
     42.9 ms  Accessors 40.04% compilation time
     22.9 ms  FunctionWrappers
   1079.6 ms  FoldsThreads 178.99% compilation time
    158.7 ms  DataStructures
      0.5 ms  SortingAlgorithms
     17.1 ms  Missings
      0.4 ms  StatsAPI
     53.0 ms  StatsBase
      5.7 ms  SimpleTraits
      1.1 ms  DelimitedFiles
     11.5 ms  MLUtils
      4.9 ms  Functors
     16.3 ms  Optimisers
      2.4 ms  GPUArraysCore
      0.1 ms  RealDot
     48.5 ms  StructArrays
     44.9 ms  ChainRules
     29.6 ms  IRTools
      0.9 ms  DiffRules
      0.3 ms  NaNMath
    375.0 ms  FillArrays
     19.0 ms  AbstractFFTs
      4.5 ms  DiffResults
   1331.8 ms  StaticArrays
      0.6 ms  CommonSubexpressions
    217.7 ms  ForwardDiff
    222.6 ms  Zygote 42.27% compilation time
     78.4 ms  OneHotArrays
     58.7 ms  Flux
 ```

</details>